+++ Model/RecordingChannel.cs
@@ -29,6 +29,7 @@ namespace DVBViewerServerApiWrapper.Model
         public bool Equals(RecordingChannel other)
         {
             if (other is null) return false;
+            if (Name == null && other.Name == null) return true;
             if (Name.Equals(other.Name)) return true;
             else return false;
         }
@@ -39,6 +40,9 @@ namespace DVBViewerServerApiWrapper.Model
         /// <returns></returns>
         public override int GetHashCode()
         {
+            if (Name == null)
+                return 0;
+
             return Name.GetHashCode();
         }
 
@@ -62,7 +66,7 @@ namespace DVBViewerServerApiWrapper.Model
             if (dvbApi != null)
             {
                 var recs = await RecordingList.GetRecordingsAsync().ConfigureAwait(false);
-                return (from f in recs.Items where f.Channel != null orderby f.Channel.Name select f.Channel).Distinct().ToList();
+                return (from f in recs.Items where f.Channel?.Name != null orderby f.Channel.Name select f.Channel).Distinct().ToList();
             }
             return null;
         }